src/runtime/pyproject.toml
[build-system]
requires = [ "setuptools>=61.0" ]
build-backend = "setuptools.build_meta"

[project]
name = "alpasim-runtime"
version = "1.46.0"
description = "Runtime of the Alpamayo Sim"
requires-python = ">=3.11,<3.12"
authors = [{name = "Michal Tyszkiewicz", email = "mtyszkiewicz@nvidia.com"}]
dependencies = [
    "aiofiles>=24.1.0",
    "alpasim-grpc",  # Local grpc package
    "alpasim-utils",  # Local utils package
    "mergedeep>=1.3.4",
    "numpy>=2.0.0",
    "omegaconf>=2.3.0",
    "prometheus-async>=25.1.0",
    "prometheus-client>=0.21.0",
    "protobuf>=4.25.3",
    "pyyaml>=6.0.2",
    "rich>=13.0.0",
    "scipy>=1.15.0",
    "csaps>=1.2.0",
    "trajdata-alpasim @ git+https://github.com/NVlabs/trajdata.git@alpasim",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "mypy>=1.0.0",
    "types-protobuf",
    "types-aiofiles",
    "grpcio>=1.70.0",
    "pytest-grpc>=0.8.0",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
nb = [
    "jupyter",
    "ipykernel",
]

[tool.uv.sources]
alpasim-grpc = {workspace = true}
alpasim-utils = {workspace = true}

[tool.setuptools.packages.find]
include = [ "alpasim_runtime", "tests" ]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "manual: marks tests as manual execution only",
]
# To run manual tests, run: `pytest -m 'manual'`
addopts = "-m 'not manual'"
